Update to BMP

+++ trunk/dev_scripts/ruby/BuildMirrorProject.rb
@@ -36,22 +36,24 @@ print 'Press any key to continue:'
 k = get_character
 puts k.chr
 
-settings   = file['Settings']
-primary    = settings['Primary']
+settings = file['Settings']
+primary = settings['Primary']
 primaryDir = settings['PrimaryDir']
-engine     = settings['Engine']
-engineDir  = settings['EngineDir']
+engine = settings['Engine']
+engineDir = settings['EngineDir']
 user = settings['User']
 
 mirror_primaryDir = File.basename(primaryDir)
-mirror_engineDir = File.basename(engineDir)
+mirror_engineDir = File.basename(engineDir) unless engineDir.nil?
 
 project_dir = File.join(current_dir, primary)
 puts '===============Creating the project directory================'
 puts "project_dir: #{project_dir}"
 FileUtils::mkdir_p project_dir
-puts "engine_dir:  #{project_dir}/#{engine}/#{mirror_engineDir}"
-FileUtils::mkdir_p "#{project_dir}/#{engine}/#{mirror_engineDir}"
+unless engineDir.nil?
+  puts "engine_dir:  #{project_dir}/#{engine}/#{mirror_engineDir}"
+  FileUtils::mkdir_p "#{project_dir}/#{engine}/#{mirror_engineDir}"
+end
 
 cyg_project_dir = `cygpath "#{project_dir}"`.chomp
 puts "cyg_project_dir: #{cyg_project_dir}"
@@ -88,10 +90,11 @@ open('pull.bat', 'w') do |io|
   orig = "#{primary}:#{primaryDir}"
   dest = "#{cyg_project_dir}/#{mirror_primaryDir}"
   io.puts "rsync -avz --exclude '.git' --exclude 'lost+found' --exclude-from '.gitignore' --delete #{user}@#{orig}/* '#{dest}'"
-
-  orig = "#{engine}:#{engineDir}"
-  dest = "#{cyg_project_dir}/#{engine}/#{mirror_engineDir}"
-  io.puts "rsync -avz --exclude '.git' --delete #{user}@#{orig}/* '#{dest}'"
+  unless engineDir.nil?
+    orig = "#{engine}:#{engineDir}"
+    dest = "#{cyg_project_dir}/#{engine}/#{mirror_engineDir}"
+    io.puts "rsync -avz --exclude '.git' --delete #{user}@#{orig}/* '#{dest}'"
+  end
 end
 
 # Create the rsync.ini file
@@ -117,8 +120,8 @@ print 'Go to cygwin and paste ssh-copy-id command then Press any key to continue
 k = get_character
 puts k.chr
 scp_file_to_unix(script_dir, '.bash_profile', user, primary)
-scp_file_to_unix(script_dir, '.aliases'     , user, primary)
-scp_file_to_unix(script_dir, '.functions'   , user, primary)
+scp_file_to_unix(script_dir, '.aliases', user, primary)
+scp_file_to_unix(script_dir, '.functions', user, primary)
 puts '==========================================================================='
 cmd = "ssh-copy-id #{user}\@#{engine}"
 puts cmd.to_s
@@ -127,8 +130,8 @@ print 'Go to cygwin and paste ssh-copy-id command then Press any key to continue
 k = get_character
 puts k.chr
 scp_file_to_unix(script_dir, '.bash_profile', user, engine)
-scp_file_to_unix(script_dir, '.aliases'     , user, engine)
-scp_file_to_unix(script_dir, '.functions'   , user, engine)
+scp_file_to_unix(script_dir, '.aliases', user, engine)
+scp_file_to_unix(script_dir, '.functions', user, engine)
 puts '==========================================================================='
 #puts "bash -c \"ssh-copy-id #{user}\@#{primary}\""
 #puts `bash -c \"ssh-copy-id #{user}\@#{primary}\"`
